Update to 0.15.1 & set buildFlags first in the command
There is one particular build flag that `go build` supports that
**must** be first in the list of flags: `-C dir`. It is used to change
into a given directory before getting started.
This flag is particularly useful when you are building a go project
inside a monorepo with a private dependency that has a replace directive
pointing to a directory inside the repository.

In our case we have two projects:
- Project A at `/project-a`
- Project B at `/project-b`

A depends on B and we want to build it using this module. If we are able
to provide `-C project-a` then we can create a directory with those two
projects and run go build.

Signed-off-by: Matias Pan <gh@matiaspan.dev>