learn the schema by studying SDL instead
TODO: validate, but the hunch is that the agent will be much better at
reading the "human intended" GraphQL schema format, which is also much
more succinct.

so, now there's a tool to learn about the schema, one type at a time. it
returns the relevant subset of the schema, which keeps the response size
small, and forces the agent to be deliberate in how it introspects.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>