doug: remove Bash, bump to Dagger v0.19.0
Trying something a bit radical: instead of giving the agent a
commandline, force it to use the available tools.

Providing Bash gives the agent too much opportunity to go off the rails,
and raises the question of what sandboxed environment the shell runs in.

Better to just implement it yourself.

Naturally, this will require the user to develop their module or other
modules to plug in. Maybe it should be easy to do that with Doug?

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>